:zap: Big improvments for multiplatform support
Adding an internal webserver to serve files in other platforms, because `file:///[...]` protocol seams not working in some other platforms in webview.
This is the only goal for adding this internal webserver.
It is secured by changing its port at each application launch (which is only in an non-admin range) and each current file is served though unique query path.

- changing default folder path from `"."` to `""`.
- improving errors handlings in Rust.
- renaming everything about `settings` into `configuration`.
- adding internal Rust webserver (Iron)
- big changes for UserData use with Arc and Mutex, motivated by webserver shared data.
- continuing to group eval calls in Rust part.
- adding `STANDALONE_MODE` in JS frontend, used for out-of-webview-debugging of HTML UI.
- voluntary ignoring preload issues because of new use of internal webserver.
- compressing base64 (by shrinking image size) of 404 default error image.
- updating Cargo dependencies (Rust).